apps/api/pyproject.toml
[build-system]
requires = ["hatchling>=1.22.0"]
build-backend = "hatchling.build"

[project]
name = "scriptkiddie-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn>=0.30.1",
  "pydantic>=2.7.4",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.1.19",
  "celery>=5.4.0",
  "redis>=5.0.6",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "httpx>=0.27.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
